Fixing touch nav top and bottom

+++ inc/gfx.h
@@ -72,14 +72,14 @@ inline uint32_t clrGetColor(const clr c)
 }
 
 //Draws text using f
-void drawText(const char *str, tex *target, const font *f, int x, int y, int sz, clr c);
+void drawText(const char *str, tex *target, const font *f, float x, float y, float sz, clr c);
 
 //Draws text wrapping lines
-void drawTextWrap(const char *str, tex *target, const font *f, int x, int y, int sz, clr c, int maxWidth);
+void drawTextWrap(const char *str, tex *target, const font *f, float x, float y, float sz, clr c, int maxWidth);
 
 //Returns text width
-size_t textGetWidth(const char *str, const font *f, int sz);
-size_t textGetHeight(const char *str, const font *f, int sz, int maxWidth);
+size_t textGetWidth(const char *str, const font *f, float sz);
+size_t textGetHeight(const char *str, const font *f, float sz, int maxWidth);
 
 //Draws rectangle at x, y with w, h
 void drawRect(tex *target, int x, int y, int w, int h, const clr c);
